more dependency cleanup
