+++ Cargo.toml
@@ -10,23 +10,14 @@ camino = "1.1.1"
 directories = "4.0.1"
 flume = { version = "0.10.14" }
 log = { version = "0.4", features = ["release_max_level_info"] }
-parking_lot = "0.12.1"
 pretty_env_logger = "0.4"
-r2d2 = "0.8.10"
-symphonia = { version = "0.5.2", features = ["mp3"] }
 thiserror = "1.0.37"
 
-iced_graphics = "0.8.0"
 iced_style = "0.8.0"
 
 clef_shared = { path = "./shared" }
 clef_db = { path = "./db" }
 clef_audio = { path = "./audio" }
-
-[dependencies.souvlaki]
-version = "0.6"
-default-features = false
-features = ["use_zbus"]
 
 [dependencies.iced]
 version = "0.9"
@@ -39,30 +30,10 @@ version = "0.10.1"
 package = "image"
 version = "0.24"
 
-[target.'cfg(target_os = "linux")'.dependencies]
-libpulse-binding = "2.5.0"
-libpulse-simple-binding = "2.5.0"
-
-[target.'cfg(not(target_os = "linux"))'.dependencies]
-cpal = "0.13.3"
-rb = "0.3.2"
-arrayvec = "0.7.1"
-rubato = "0.12.0"
-
-[target.'cfg(target_os = "windows")'.dependencies]
-once_cell = "1.17.1"
-
-[target.'cfg(target_os = "windows")'.dependencies.windows]
-version = "0.48.0"
-features = [
-    "Win32_Foundation",
-    "Win32_UI_WindowsAndMessaging"
-]
-
 [dev-dependencies]
 mockall = "0.11.3"
 
-# compile all dependencies in release mode
+# compile all external dependencies in release mode
 [profile.dev.package."*"]
 opt-level = 3
