+++ audio/src/lib.rs
@@ -1,5 +1,10 @@
+#![warn(rust_2018_idioms)]
+#![deny(missing_debug_implementations)]
+#![deny(unsafe_code)]
+
+pub mod metadata;
 pub mod player;
-pub use player::track_info::{first_supported_track, TrackInfo};
+pub mod track_info;
 
 #[cfg(not(target_os = "linux"))]
 mod resampler;
