+++ db/Cargo.toml
@@ -8,7 +8,6 @@ authors = [ "Dan Knutson <dan.knutson@gmail.com>" ]
 camino = "1.1.1"
 diesel_migrations = { version = "2.0.0", features = ["sqlite"] }
 log = { version = "0.4", features = ["release_max_level_info"] }
-parking_lot = "0.12.1"
 r2d2 = "0.8.10"
 thiserror = "1.0.37"
 libsqlite3-sys = { version = "0.25.2", features = ["bundled"] }
