+++ shared/src/lib.rs
@@ -1,3 +1,7 @@
+#![warn(rust_2018_idioms)]
+#![deny(missing_debug_implementations)]
+#![deny(unsafe_code)]
+
 use std::collections::VecDeque;
 use std::fmt::Debug;
